refactor(navbar): removing dynamic avatars from navbar
